Refactor: moved StudentExpel`s methods to StudentExpelController
